+++ lib/airbrake-ruby/config.rb
@@ -103,7 +103,7 @@ module Airbrake
       @endpoint ||=
         begin
           self.host = ('https://' << host) if host !~ %r{\Ahttps?://}
-          api = "/api/v3/projects/#{project_id}/notices?key=#{project_key}"
+          api = "api/v3/projects/#{project_id}/notices?key=#{project_key}"
           URI.join(host, api)
         end
     end
@@ -162,11 +162,6 @@ module Airbrake
       __send__("#{option}=", value)
     rescue NoMethodError
       raise Airbrake::Error, "unknown option '#{option}'"
-    end
-
-    def set_endpoint(id, key, host)
-      host = ('https://' << host) if host !~ %r{\Ahttps?://}
-      @endpoint = URI.join(host, "/api/v3/projects/#{id}/notices?key=#{key}")
     end
   end
 end